apply changes
