+++ src/common/fixtures.ts
@@ -1,0 +1,148 @@
+export const employment = {
+  data: {
+    employment: {
+      pricing_plan_details: {
+        frequency: 'monthly',
+      },
+      administrative_details: null,
+      manager_id: null,
+      billing_address_details: null,
+      work_email: 'joao.almeida+amendmentsdk@remote.com',
+      personal_details: null,
+      user_id: '01e05d10-bcdd-4b4f-a52e-7a5c41f7381a',
+      seniority_date: null,
+      updated_at: '2025-04-11T09:14:17',
+      emergency_contact_details: null,
+      job_title: 'engineer',
+      manager_employment_id: null,
+      contract_details: {
+        annual_gross_salary: 35000000,
+        annual_training_hours_ack: 'acknowledged',
+        available_pto: 22,
+        available_pto_type: 'unlimited',
+        benefits: {
+          business_travel_insurance:
+            'Basic Business Travel (Expatriate Group - Basic Business Travel)',
+          health_insurance_2025:
+            'Basic Health Plan 2025 (Single) (Advance Care/ Tranquilidade - Basic Health Plan 2025 (Single))',
+          life_insurance:
+            'Life Insurance - $50K (Allianz Life - Global Life - Life 50k)',
+          meal_benefit:
+            'Meal Allowance Standard 2025 (Meal Allowance - Meal Allowance Standard 2025)',
+          mental_health_program:
+            'Basic Mental Health Program (Oliva - Basic Mental Health (Platform Access Only))',
+        },
+        contract_duration_type: 'indefinite',
+        equity_compensation: {
+          offer_equity_compensation: 'no',
+        },
+        experience_level:
+          'Level 3 - Associate - Employees who perform independently tasks and/or with coordination and control functions',
+        has_bonus: 'no',
+        has_commissions: 'no',
+        has_signing_bonus: 'no',
+        probation_length_days: 40,
+        role_description:
+          'Very nice job.Very nice job.Very nice job.Very nice job.Very nice job.Very nice job.Very nice job.Very nice job.',
+        salary_installments_confirmation: 'acknowledged',
+        work_address: {
+          is_home_address: 'yes',
+        },
+        work_from_home_allowance_ack: 'acknowledged',
+        work_hours_per_week: 40,
+        work_schedule: 'full_time',
+        working_hours_exemption: 'no',
+      },
+      department: null,
+      personal_email: 'joao.almeida+amendmentsdk@remote.com',
+      active_contract_id: 'b6577c50-43e0-4fb4-9264-7442d4659653',
+      termination_date: null,
+      manager: null,
+      probation_period_end_date: '2025-06-08',
+      provisional_start_date: '2025-04-30',
+      status: 'created',
+      country: {
+        code: 'PRT',
+        name: 'Portugal',
+        alpha_2_code: 'PT',
+        supported_json_schemas: [
+          'additional_documents',
+          'administrative_details',
+          'contract_details',
+          'employment_basic_information',
+          'emergency_contact',
+          'address_details',
+        ],
+      },
+      created_at: '2025-04-11T09:12:46',
+      company_id: 'c3c22940-e118-425c-9e31-f2fd4d43c6d8',
+      department_id: null,
+      external_id: null,
+      short_id: 'S079JG',
+      basic_information: {
+        name: 'Joao Almeida',
+        email: 'joao.almeida+amendmentsdk@remote.com',
+        provisional_start_date: '2025-04-30',
+        job_title: 'engineer',
+        tax_servicing_countries: [],
+        has_seniority_date: 'no',
+        tax_job_category: 'engineering_it',
+      },
+      address_details: null,
+      files: [],
+      eligible_for_onboarding_cancellation: true,
+      id: '2ef4068b-11c7-4942-bb3c-70606c83688e',
+      full_name: 'Joao Almeida',
+      employment_lifecycle_stage: 'employment_creation',
+      type: 'employee',
+      bank_account_details: [],
+      manager_email: null,
+      onboarding_tasks: {
+        bank_account_details: {
+          status: 'pending',
+          description: 'Bank account used for receiving salary payments.',
+        },
+        administrative_details: {
+          status: 'pending',
+          description: 'Information we need for tax purposes.',
+        },
+        address_details: {
+          status: 'pending',
+          description: 'Primary residence.',
+        },
+        contract_details: {
+          status: 'completed',
+          description:
+            'Employee-specific details for their employment agreement.',
+        },
+        personal_details: {
+          status: 'pending',
+          description: 'Personal details, such as name and date of birth.',
+        },
+        emergency_contact_details: {
+          status: 'pending',
+          description: 'Who should be called in an emergency.',
+        },
+        pricing_plan_details: {
+          status: 'completed',
+          description:
+            'How often Remote will bill employers for management fees.',
+        },
+        billing_address_details: {
+          status: 'pending',
+          description: "Address associated with the employee's bank account.",
+        },
+        employment_document_details: {
+          status: 'pending',
+          description: 'We need some additional documents.',
+        },
+        employment_eligibility: {
+          status: 'pending',
+          description:
+            'We’ll make sure you can work in the country where you live.',
+        },
+      },
+      user_status: 'created',
+    },
+  },
+};
